+++ generator_tests/test/freezed/freezed.dart
@@ -7,7 +7,9 @@ part 'freezed.freezed.dart';
 @freezed
 @ReactiveFormAnnotation()
 class Test with _$Test {
-  const factory Test(
+  Test._();
+
+  factory Test(
     @FormControlAnnotation() String? gender, {
     @FormControlAnnotation() String? id,
     @FormControlAnnotation() String? name,